+++ src/ffi/lapack.rs
@@ -23,3 +23,21 @@ unsafe extern "C" {
         info: *mut __CLPK_integer,
     );
 }
+
+#[allow(
+    missing_docs,
+    non_camel_case_types,
+    non_snake_case,
+    non_upper_case_globals,
+    unused_imports,
+    dead_code,
+    improper_ctypes,
+    improper_ctypes_definitions,
+    unnecessary_transmutes
+)]
+mod generated {
+    use super::*;
+    include!("generated/lapack_missing.rs");
+}
+
+pub use generated::*;