+++ src/ffi/mod.rs
@@ -1,4 +1,4 @@
-#![allow(dead_code, unused_imports)]
+#![allow(ambiguous_glob_reexports, dead_code, unused_imports)]
 
 pub mod blas;
 pub mod bnns;
@@ -7,6 +7,7 @@ pub mod quadrature;
 pub mod simd;
 pub mod sparse;
 pub mod vdsp;
+pub mod veclib_extras;
 pub mod vforce;
 pub mod vimage;
 
@@ -17,5 +18,6 @@ pub use self::quadrature::*;
 pub use self::simd::*;
 pub use self::sparse::*;
 pub use self::vdsp::*;
+pub use self::veclib_extras::*;
 pub use self::vforce::*;
 pub use self::vimage::*;