+++ src/ffi/vimage.rs
@@ -84,3 +84,21 @@ unsafe extern "C" {
         flags: vImage_Flags,
     ) -> vImage_Error;
 }
+
+#[allow(
+    missing_docs,
+    non_camel_case_types,
+    non_snake_case,
+    non_upper_case_globals,
+    unused_imports,
+    dead_code,
+    improper_ctypes,
+    improper_ctypes_definitions,
+    unnecessary_transmutes
+)]
+mod generated {
+    use super::*;
+    include!("generated/vimage_missing.rs");
+}
+
+pub use generated::*;